refactor: config_lib, remove unneed part
